server/src/test/java/org/mitre/jwt/signer/service/impl/KeyStoreTest.java
package org.mitre.jwt.signer.service.impl;

import static org.hamcrest.CoreMatchers.not;
import static org.hamcrest.CoreMatchers.nullValue;
import static org.junit.Assert.assertThat;

import java.io.IOException;
import java.security.GeneralSecurityException;
import java.security.Key;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

@SuppressWarnings("restriction") // I know... 
@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(locations = {
		"classpath:test-context.xml" })
public class KeyStoreTest {

	@Autowired
	@Qualifier("testKeystore")
	KeyStore keystore;

	@Test
	public void storeKeyPair() throws GeneralSecurityException, IOException {
//
//		java.security.KeyStore ks = KeyStore.generateRsaKeyPair(keystore
//				.getLocation().getFile().getPath(), "OpenID Connect Server",
//				"test", KeyStore.PASSWORD, KeyStore.PASSWORD, 30, 30);
//
//		keystore.setKeystore(ks);
//
//		assertThat(ks, not(nullValue()));
		assertThat(true, not(false));
	}

	@Test
	public void readKey() throws GeneralSecurityException {

//		Key key = keystore.getKeystore().getKey("test",
//				KeyStore.PASSWORD.toCharArray());
//
//		System.out.println("-----BEGIN PRIVATE KEY-----");
//		System.out
//				.println(new sun.misc.BASE64Encoder().encode(key.getEncoded()));
//		System.out.println("-----END PRIVATE KEY-----");
//
//		assertThat(key, not(nullValue()));
		assertThat(true, not(false));
	}
}
